Fixed input data day offset
This PR fixes an issue with processing lateral flow input files when the file pattern was "nex-" or "cat-".

Problem

Previously, the start-time parameter — which defines the beginning of the data simulation — was not being properly applied when reading "nex-" or "cat-" input files. As a result, the simulation ran over an incorrect time window.

For example:
If the lateral flow input covers 2023-04-01 to 2023-04-03, nts = 24, and the start time is 2023-04-02_00:00, the simulation should run from 2023-04-02_00:00 to 2023-04-02_23:00.
However, the previous implementation simulated from 2023-04-01_00:00 to 2023-04-02_00:23, which was incorrect.

Fix

The updated logic now correctly accounts for the specified start time when reading "nex-" or "cat-" files, ensuring that the simulation window aligns with the intended period.

I used a tool to make the code pythonic and it looks like that changed the code heavily. But the real change is in build_qlateral_array() function. Everything else is just formatting related change, which can be completely ignored.

Authored by: Sonam Lama slama@ua.edu